Cleanup protocol file
